Changed long to double for token values

+++ AdjencyList.java
@@ -1,15 +1,15 @@
 import java.util.*;
 public class AdjencyList{
     private String baseAddress;
-    private HashMap<String, Long> adjencyList;
+    private HashMap<String, Double> adjencyList;
 
-    public AdjencyList(String baseAddress, String toAddress, long token){
+    public AdjencyList(String baseAddress, String toAddress, double token){
         this.baseAddress = baseAddress;
-        adjencyList = new HashMap<String, Long>();
+        adjencyList = new HashMap<String, Double>();
         adjencyList.put(toAddress, token);
     }
 
-    public HashMap<String, Long> getAdjencyList() {
+    public HashMap<String, Double> getAdjencyList() {
         return adjencyList;
     }
